# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Knowledge Melting Pot API
  version: 1.0.0
  description: Versioned API for auditable knowledge extraction and publication.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /api/v1
security:
  - sessionCookie: []
paths:
...
  /subscenes/{subSceneId}/extraction-jobs:
    post:
      summary: Start a knowledge extraction job
      operationId: startExtraction
      parameters:
        - $ref: '#/components/parameters/SubSceneId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartExtractionRequest'
      responses:
        '202':
          $ref: '#/components/responses/AcceptedJob'
        '400':
          $ref: '#/components/responses/Problem'
...
  /jobs/{jobId}:
    get:
      summary: Get durable job state
      operationId: getJob
      parameters:
        - $ref: '#/components/parameters/JobId'
      responses:
        '200':
          description: Durable job state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '404':
          $ref: '#/components/responses/Problem'
  /jobs/{jobId}/events:
    get:
      summary: Stream ordered replayable job events
      operationId: streamJobEvents
      parameters:
        - $ref: '#/components/parameters/JobId'
        - name: Last-Event-ID
          in: header
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Ordered and replayable job events.
          content:
            text/event-stream:
              schema:
                $ref: '#/components/schemas/JobEventEnvelope'
                description: >-
                  SSE frames named stage-started, progress, preview, warning, completed, or failed.
                  Every data field is a serialized JobEventEnvelope and Last-Event-ID is the numeric sequence.
        '404':
          $ref: '#/components/responses/Problem'
  /jobs/{jobId}/cancel:
    post:
      summary: Request cooperative job cancellation
      operationId: cancelJob
      parameters:
        - $ref: '#/components/parameters/JobId'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '202':
          $ref: '#/components/responses/AcceptedJob'
        '404':
          $ref: '#/components/responses/Problem'
  /jobs/{jobId}/retry:
    post:
      summary: Retry a failed job
      operationId: retryJob
      parameters:
        - $ref: '#/components/parameters/JobId'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '202':
          $ref: '#/components/responses/AcceptedJob'
        '404':
          $ref: '#/components/responses/Problem'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: KMP_SESSION
  parameters:
...
    SubSceneId:
      name: subSceneId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    JobId:
      name: jobId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
        minLength: 8
        maxLength: 128
...
  responses:
    AcceptedJob:
      description: Command accepted as a durable job.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AcceptedJob'
    Problem:
      description: RFC 7807 problem response.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
...
    StartExtractionRequest:
      type: object
      additionalProperties: false
      required: [roundId, modelConfigVersionId, skillVersionId]
      properties:
        roundId: { type: string, format: uuid }
        modelConfigVersionId: { type: string, format: uuid }
        skillVersionId: { type: string, format: uuid }
...
    AcceptedJob:
      type: object
      required: [jobId, status, statusUrl, eventsUrl]
      properties:
        jobId: { type: string, format: uuid }
        status:
          type: string
          enum: [QUEUED, RUNNING, SUCCEEDED, FAILED, CANCELLED]
        statusUrl: { type: string }
        eventsUrl: { type: string }
        resourceType: { type: [string, 'null'] }
        resourceId: { type: [string, 'null'], format: uuid }
    Job:
      type: object
      required: [id, type, status, attempt, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        type:
          type: string
          enum: [INGEST, EXTRACT, REEXTRACT, ALIGN, GENERATE_ASSET, GENERATE_ALL, SCENE_EXPLORE, EVALUATE]
        status:
          type: string
          enum: [QUEUED, RUNNING, SUCCEEDED, FAILED, CANCELLED]
        stage: { type: [string, 'null'] }
        percent: { type: integer, minimum: 0, maximum: 100 }
        attempt: { type: integer, minimum: 0 }
        errorCode: { type: [string, 'null'] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    JobEventEnvelope:
      type: object
      additionalProperties: false
      required: [eventId, sequence, jobId, stage, percent, messageCode, traceId, timestamp, message]
      properties:
        eventId: { type: string, pattern: '^[0-9]+$' }
        sequence: { type: integer, format: int64, minimum: 0 }
        jobId: { type: string, format: uuid }
        stage: { type: string, pattern: '^[A-Za-z0-9_-]{1,64}$' }
        percent: { type: integer, minimum: 0, maximum: 100 }
        messageCode:
          type: string
          pattern: '^[A-Za-z0-9._:-]{1,128}$'
          description: >
            Stable machine-readable code. Ingest jobs emit the diagnostic
            EMBEDDING_PROVIDER_UNCONFIGURED when no real embedding provider is
            configured; vectors are never fabricated and no EMBEDDINGS_COMMITTED
            stage is reported in that state.
        traceId: { type: string, pattern: '^[A-Za-z0-9._:-]{1,128}$' }
        timestamp: { type: string, format: date-time }
        message: { type: string }
...
    Problem:
      type: object
      required: [type, title, status, code, traceId]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        detail: { type: string }
        instance: { type: string, format: uri-reference }
        code: { type: string }
        traceId: { type: string }
        errors:
          type: array
          items:
            type: object
            required: [field, message]
            properties:
              field: { type: string }
              message: { type: string }